Split the module into subfolder
